refactor(settings): split the settings definitions into section modules
settingsDefinitions.ts had grown to 2409 lines holding twenty-three page
and group builders. None of them called another, so understanding one row
meant paging through a dozen unrelated sections and a search by setting
name matched in three roles at once.

Each builder now describes its own section in a module under
src/settings/sections, named for the section it draws. What they share
moved to three modules they all import and none of them import back: the
context and the block classes, the row shapes more than one section needs,
and the control keys that address a list entry rather than a property.
What a control key means once a value is read or written turned out to be
a separate question from what the tree looks like, so it moved to
src/settings/controlValues.ts.

settingsDefinitions.ts is left with the ordering of the sections and
re-exports, at 173 lines. SettingsTab and legacySettingsRenderer import
the same paths as before and are untouched.

No row, alias, order, description, visibility or disabled predicate
changed. tests/integration/settingsDefinitions.test.ts passes unmodified,
including the invariants that walk the whole tree and would catch a row
that went missing.
